app/user_service/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
PyJWT
aiokafka==0.8.1
